Split Violations by category chart and card components.
